About Page BG
* Introduce interactive hex bg
* Move screensize store and enhance solution
