+++ packages/frontend/tsconfig.json
@@ -10,8 +10,4 @@
 		"sourceMap": true,
 		"strict": true
 	}
-	// Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
-	//
-	// If you want to overwrite includes/excludes, make sure to copy over the relevant includes/excludes
-	// from the referenced tsconfig.json - TypeScript does not merge them in
 }
